+++ Road.cc
@@ -5,7 +5,7 @@
 // @param destination - Pointer to the Town this road leads to
 // @param distance   - Length/weight of the road (in arbitrary units)
 // @param isBridge   - Boolean flag indicating if this road is a bridge
-Road::Road(Town* destination, int distance, bool isBridge)
+Road::Road(Town* destination, float distance, bool isBridge)
     : destination(destination), distance(distance), isBridge(isBridge) {}
 
 // Getter method for the destination town
@@ -14,7 +14,7 @@ Town* Road::getDestination() const {
 }
 
 // Getter method for the road's distance/length
-int Road::getDistance() const {
+float Road::getDistance() const {
     return distance;
 }
